requirements.txt
pdfminer.six==20221105
streamlit==1.31.1
